Define the biconditional connective.
Also clean up the coverage script and include metavariables.
Metamath-style definitions can only define logical constants; in order
to express a variable, we must treat it as a sort of floating hole which
weaves through logical structure, and this means that it is declared $f
instead of $a. So, I wrote a regex to match $f and another to match $a.

This definition of the biconditional is almost directly from iset.mm,
and I think it's worth noting that I only prefer it because it allows
for reduction of axioms. In general, I want few axioms, because I want
the amount of certainty to be maximized; I'm willing to eat a few long
{ganai} and {ge} and {go} in order to get there.