Define the lattice of relations.
Two of the definitions, ~df-kihirvlina and ~df-kihirkanxe, are
interesting; they don't use a {gi}, but state that all of their related
terbri will always together obey some *syntactic* property. This is
truly remarkable, as it suggests an invariance which could be used more
fundamentally.

Lemme expand on that; I might return to this later. If a selbri is
completed with terbri composed entirely of free metavariables, and is
not just bridi but |- inhabited, then this makes it part of a
pre-Lojbanic skeleton. This skeleton could be shown to axiomatically
fulfill the requirements of a particular categorical structure, e.g. a
bicategory of relations, prior to the construction of categorical logic.

In particular, we could ground {ge} and {ga} in {je} and {ja}, which is
backwards from how it's currently done, and also ground {je} and {ja} in
{ki'irkanxe} and {ki'irvlina}. This...has no obvious practical benefit.