Bump zaha for spati DAG.
I know that keeping these repositories separate might be the wrong
choice in the long run (should they live here?) but right now it is
performing the valuable job of forcing me to only draw diagrams
including the arrows which I'm most confident are correct.

I'd rather be conservative than wrong here. As we climb the ontological
ladder from the foundations up to our desired sky-castles, we will have
to be increasingly careful with powerful absolute statements.
Fortunately, subrelations are both uncommon and easy to check.