I'm checking in another unproved theorem.
I had hoped that the recent buffs to universal quantification were
enough to let me fill out ~du-kinfi. However, I can't quite finish the
proof. The problem is that ~du-sym is not in deductive form, but
inference form, and ~du-kinfi is a closed deduction. I'm not going to
explain all of that here, but it's documented:
https://us.metamath.org/mpeuni/mmnatded.html

I might be able to overcome the issue by going back and reproving a
bunch of stuff in deductive form. However, I'm a little tired right now
and don't want to do that unless I can understand how it will look.