Basic definition of {gripau}.
Fuck it, I'm giving up and treating {gripau} as an irregular lujvo. Its
definition has nothing whatsoever to do with its rafsi -gri- and -pau-.
When set up this way, we should be able to inherit properties of
{ganai}, like transitivity, but I need to add a few theorems for
manipulating {na.a} first, and those depend upon manipulating {se}.

Let me spell it out in more detail, just in case:
* Theorems like {ganai ko'a bu'a ko'e gi ko'e se bu'a ko'a}
* Can then deduce something like
  {ganai ko'a na.a ko'e bu'a ko'i gi ko'i se bu'a ko'a na.a ko'e}
* And then finally we have sugar for {na.a} which lifts to {gripau}

See {du} for a worked example of this sort of setup. There, we needed to
work under {ro bu'a zo'u}, requiring similar helpers.