Rewrite intro to nats to do VUhU operators first.
Appears to give a simpler and easier-to-manipulate version of successor
axioms, although it comes at the cost of always having a {li} head. But
maybe that's an acceptable impedance mismatch since it will force us to
use objects which are structually numbers (either literal or operator)
rather than guarding that {ko'a} is a nat.