Add an interesting paper to the bibliography.
I am going to have to chew on this paper's ideas for a bit. The central
idea is to use their system FPA to prove *conditional* results about the
natural numbers without assuming their existence. For example, instead
of *assuming* that zero is a nat, *prove* that *if* there is at least
one nat, *then* zero is a nat.

There are obvious downsides to FPA, but Boucher proposes it primarily as
a way to evade Gödel's incompleteness theorems by sacrificing something
new: the ability to assume all natural numbers exist! Instead FPA says
that it is complete over all initial segments of the natural numbers,
each of which gives a model, but cannot be complete over the natural
numbers themselves.

Since folks have accused me of going one way or the other, let me state
up front that I think that FPA should be thought of as powerful evidence
that PA (and Z₂, second-order PA) are consistent. It is already
well-established without Boucher's paper that all (powers of) primes
give models of modular arithmetic, which is also a way of phrasing PA
without the successor axiom. For any particular ceiling, we may define a
finite fragment of PA which assumes nothing beyond that ceiling, and we
may be confident of that fragment's consistency. At the same time, we
won't directly encourage working without the successor axiom more than
we already encourage removal of any unnecessary axioms.