Set up a little framework for reflexivity.
The idea is that, for our selbri of interest, we first prove a
standalone theorem like:

    |- ko'a bu'a ko'a

Which usually can be proved from {ganai}, {go}, or relatives. Since
those connectives are the most primitive in the system, these theorems
are genuinely universal in their metasyntactic quantification. We can
therefore restrict them over any domain whatsoever, and as long as the
caller is careful to keep {ko'e} as a set, we can then prove:

    |- pa ka ce'u bu'a ce'u kei kinra ko'e

And this inference is formalized as ~refl-kinra. Examples include
~du-refl, ~gripau-refl, and ~kihirnihi-refl, which all have
corresponding theorems ~du-kinra, ~gripau-kinra, and ~kihirnihi-kinra
generated by composition of inferences.